Seperated the API call functions and csv functions into separate files and added required library to requirements.txt
